changed to unsigned long for c++ complience

+++ cpp_projs/CPP06/ex01/Serialization.hpp
@@ -12,7 +12,7 @@ class Serialization {
         ~Serialization();
     public:
         static uintptr_t serialize(Data *ptr);
-        static Data* deserialize(uintptr_t raw);
+        static Data* deserialize(unsigned long raw);
 };
 
 #endif
